+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 ### New
 - Added a system for per-player text tones. Text tones allow you to set a custom sound that replaces the sound used for speech or private comms on an individual basis for each player. If you wish to add new sounds, create an "alternate" folder inside sounds/ and add them there; the in-client browser will see them. They must be formatted as Ogg Vorbis (.ogg). Type 'conf tones' to begin.
+- Text tones now apply to OOC, SOOC, and ROOC. The communicator tone plays on top of the normal OOC sound instead of replacing it.
 
 ## Version 4.2.13